Unify authorization testing
