+++ work-item-link-category-blackbox_test.go
@@ -2,10 +2,8 @@ package main_test
 
 import (
 	"bytes"
-	"encoding/json"
 	"fmt"
 	"net/http"
-	"net/http/httptest"
 	"testing"
 
 	. "github.com/almighty/almighty-core"
@@ -17,8 +15,6 @@ import (
 	"github.com/almighty/almighty-core/resource"
 	jwt "github.com/dgrijalva/jwt-go"
 	"github.com/goadesign/goa"
-	"github.com/goadesign/goa/middleware"
-	goajwt "github.com/goadesign/goa/middleware/security/jwt"
 	"github.com/jinzhu/gorm"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
@@ -183,7 +179,7 @@ func (s *WorkItemLinkCategorySuite) TestListWorkItemLinkCategory() {
 	toBeFound := 2
 	for i := 0; i < len(linkCatCollection.Data) && toBeFound > 0; i++ {
 		if *linkCatCollection.Data[i].Data.Attributes.Name == "system" || *linkCatCollection.Data[i].Data.Attributes.Name == "user" {
-			s.T().Log("Found work item link category in collection: ", linkCatCollection.Data[i].Data.Attributes.Name)
+			s.T().Log("Found work item link category in collection: ", *linkCatCollection.Data[i].Data.Attributes.Name)
 			toBeFound--
 		}
 	}
@@ -195,24 +191,6 @@ func (s *WorkItemLinkCategorySuite) TestListWorkItemLinkCategory() {
 func TestSuiteWorkItemLinkCategory(t *testing.T) {
 	resource.Require(t, resource.Database)
 	suite.Run(t, new(WorkItemLinkCategorySuite))
-}
-
-// Expected strcture of 401 error response
-type errorResponseStruct struct {
-	Id     string
-	Code   string
-	Status int
-	Detail string
-}
-
-// testSecureAPI defines how a Test object is.
-type testSecureAPI struct {
-	method             string
-	url                string
-	expectedStatusCode int    // this will be tested against responseRecorder.Code
-	expectedErrorCode  string // this will be tested only if response body gets unmarshelled into errorResponseStruct
-	payload            *bytes.Buffer
-	jwtToken           string
 }
 
 func getWorkItemLinkCategoryTestData(t *testing.T) []testSecureAPI {
@@ -243,28 +221,28 @@ func getWorkItemLinkCategoryTestData(t *testing.T) []testSecureAPI {
 		// Create Work Item API with different parameters
 		{
 			method:             http.MethodPost,
-			url:                EndpointWorkItemLinkCategories,
+			url:                endpointWorkItemLinkCategories,
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            createWorkItemLinkCategoryPayloadString,
-			jwtToken:           GetExpiredAuthHeader(t, privatekey),
+			jwtToken:           getExpiredAuthHeader(t, privatekey),
 		}, {
 			method:             http.MethodPost,
-			url:                EndpointWorkItemLinkCategories,
+			url:                endpointWorkItemLinkCategories,
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            createWorkItemLinkCategoryPayloadString,
-			jwtToken:           GetMalformedAuthHeader(t, privatekey),
+			jwtToken:           getMalformedAuthHeader(t, privatekey),
 		}, {
 			method:             http.MethodPost,
-			url:                EndpointWorkItemLinkCategories,
+			url:                endpointWorkItemLinkCategories,
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            createWorkItemLinkCategoryPayloadString,
-			jwtToken:           GetValidAuthHeader(t, differentPrivatekey),
+			jwtToken:           getValidAuthHeader(t, differentPrivatekey),
 		}, {
 			method:             http.MethodPost,
-			url:                EndpointWorkItemLinkCategories,
+			url:                endpointWorkItemLinkCategories,
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            createWorkItemLinkCategoryPayloadString,
@@ -273,28 +251,28 @@ func getWorkItemLinkCategoryTestData(t *testing.T) []testSecureAPI {
 		// Update Work Item API with different parameters
 		{
 			method:             http.MethodPut,
-			url:                EndpointWorkItemLinkCategories + "/6c5610be-30b2-4880-9fec-81e4f8e4fd76",
+			url:                endpointWorkItemLinkCategories + "/6c5610be-30b2-4880-9fec-81e4f8e4fd76",
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            createWorkItemLinkCategoryPayloadString,
-			jwtToken:           GetExpiredAuthHeader(t, privatekey),
+			jwtToken:           getExpiredAuthHeader(t, privatekey),
 		}, {
 			method:             http.MethodPut,
-			url:                EndpointWorkItemLinkCategories + "/6c5610be-30b2-4880-9fec-81e4f8e4fd76",
+			url:                endpointWorkItemLinkCategories + "/6c5610be-30b2-4880-9fec-81e4f8e4fd76",
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            createWorkItemLinkCategoryPayloadString,
-			jwtToken:           GetMalformedAuthHeader(t, privatekey),
+			jwtToken:           getMalformedAuthHeader(t, privatekey),
 		}, {
 			method:             http.MethodPut,
-			url:                EndpointWorkItemLinkCategories + "/6c5610be-30b2-4880-9fec-81e4f8e4fd76",
+			url:                endpointWorkItemLinkCategories + "/6c5610be-30b2-4880-9fec-81e4f8e4fd76",
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            createWorkItemLinkCategoryPayloadString,
-			jwtToken:           GetValidAuthHeader(t, differentPrivatekey),
+			jwtToken:           getValidAuthHeader(t, differentPrivatekey),
 		}, {
 			method:             http.MethodPut,
-			url:                EndpointWorkItemLinkCategories + "/6c5610be-30b2-4880-9fec-81e4f8e4fd76",
+			url:                endpointWorkItemLinkCategories + "/6c5610be-30b2-4880-9fec-81e4f8e4fd76",
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            createWorkItemLinkCategoryPayloadString,
@@ -303,28 +281,28 @@ func getWorkItemLinkCategoryTestData(t *testing.T) []testSecureAPI {
 		// Delete Work Item API with different parameters
 		{
 			method:             http.MethodDelete,
-			url:                EndpointWorkItemLinkCategories + "/6c5610be-30b2-4880-9fec-81e4f8e4fd76",
+			url:                endpointWorkItemLinkCategories + "/6c5610be-30b2-4880-9fec-81e4f8e4fd76",
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            nil,
-			jwtToken:           GetExpiredAuthHeader(t, privatekey),
+			jwtToken:           getExpiredAuthHeader(t, privatekey),
 		}, {
 			method:             http.MethodDelete,
-			url:                EndpointWorkItemLinkCategories + "/6c5610be-30b2-4880-9fec-81e4f8e4fd76",
+			url:                endpointWorkItemLinkCategories + "/6c5610be-30b2-4880-9fec-81e4f8e4fd76",
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            nil,
-			jwtToken:           GetMalformedAuthHeader(t, privatekey),
+			jwtToken:           getMalformedAuthHeader(t, privatekey),
 		}, {
 			method:             http.MethodDelete,
-			url:                EndpointWorkItemLinkCategories + "/6c5610be-30b2-4880-9fec-81e4f8e4fd76",
+			url:                endpointWorkItemLinkCategories + "/6c5610be-30b2-4880-9fec-81e4f8e4fd76",
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            nil,
-			jwtToken:           GetValidAuthHeader(t, differentPrivatekey),
+			jwtToken:           getValidAuthHeader(t, differentPrivatekey),
 		}, {
 			method:             http.MethodDelete,
-			url:                EndpointWorkItemLinkCategories + "/6c5610be-30b2-4880-9fec-81e4f8e4fd76",
+			url:                endpointWorkItemLinkCategories + "/6c5610be-30b2-4880-9fec-81e4f8e4fd76",
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            nil,
@@ -334,7 +312,7 @@ func getWorkItemLinkCategoryTestData(t *testing.T) []testSecureAPI {
 		// We do not have security on GET hence this should return 404 not found
 		{
 			method:             http.MethodGet,
-			url:                EndpointWorkItemLinkCategories + "/fc591f38-a805-4abd-bfce-2460e49d8cc4",
+			url:                endpointWorkItemLinkCategories + "/fc591f38-a805-4abd-bfce-2460e49d8cc4",
 			expectedStatusCode: http.StatusNotFound,
 			expectedErrorCode:  "not_found",
 			payload:            nil,
@@ -345,69 +323,11 @@ func getWorkItemLinkCategoryTestData(t *testing.T) []testSecureAPI {
 
 // This test case will check authorized access to Create/Update/Delete APIs
 func TestUnauthorizeWorkItemLinkCategoryCUD(t *testing.T) {
-	resource.Require(t, resource.Database)
-
-	// This will be modified after merge PR for "Viper Environment configurations"
-	publickey, err := jwt.ParseRSAPublicKeyFromPEM((configuration.GetTokenPublicKey()))
-	if err != nil {
-		t.Fatal("Could not parse Key ", err)
-	}
-	tokenTests := getWorkItemLinkCategoryTestData(t)
-
-	for _, testObject := range tokenTests {
-		// Build a request
-		var req *http.Request
-		var err error
-		if testObject.payload == nil {
-			req, err = http.NewRequest(testObject.method, testObject.url, nil)
-		} else {
-			req, err = http.NewRequest(testObject.method, testObject.url, testObject.payload)
-		}
-		// req, err := http.NewRequest(testObject.method, testObject.url, testObject.payload)
-		if err != nil {
-			t.Fatal("could not create a HTTP request")
-		}
-		// Add Authorization Header
-		req.Header.Add("Authorization", testObject.jwtToken)
-
-		rr := httptest.NewRecorder()
-
-		// temperory service for testing the middleware
-		service := goa.New("TestUnauthorizedCreateWorkItemLinkCategory-Service")
-		assert.NotNil(t, service)
-
-		// if error is thrown during request processing, it will be caught by ErrorHandler middleware
-		// this will put error code, status, details in recorder object.
-		// e.g> {"id":"AL6spYb2","code":"jwt_security_error","status":401,"detail":"JWT validation failed: crypto/rsa: verification error"}
-		service.Use(middleware.ErrorHandler(service, true))
-
-		// append a middleware to service. Use appropriate RSA keys
-		jwtMiddleware := goajwt.New(publickey, nil, app.NewJWTSecurity())
-		// Adding middleware via "app" is important
-		// Because it will check the design and accordingly apply the middleware if mentioned in design
-		// But if I use `service.Use(jwtMiddleware)` then middleware is applied for all the requests (without checking design)
-		app.UseJWTMiddleware(service, jwtMiddleware)
-
-		controller := NewWorkitemController(service, gormapplication.NewGormDB(DB))
-		app.MountWorkitemController(service, controller)
-
-		// Hit the service with own request
-		service.Mux.ServeHTTP(rr, req)
-
-		assert.Equal(t, testObject.expectedStatusCode, rr.Code)
-
-		// Below code tries to open Body response which is expected to be a JSON
-		// If could not parse it correctly into errorResponseStruct
-		// Then it gets logged and continue the test loop
-		content := new(errorResponseStruct)
-		err = json.Unmarshal(rr.Body.Bytes(), content)
-		if err != nil {
-			t.Log("Could not parse JSON response: ", rr.Body)
-			// safe to continue because we alread checked rr.Code=required_value
-			continue
-		}
-		// Additional checks for 'more' confirmation
-		assert.Equal(t, testObject.expectedErrorCode, content.Code)
-		assert.Equal(t, testObject.expectedStatusCode, content.Status)
-	}
+	UnauthorizeCreateUpdateDeleteTest(t, getWorkItemLinkCategoryTestData, func() *goa.Service {
+		return goa.New("TestUnauthorizedCreateWorkItemLinkCategory-Service")
+	}, func(service *goa.Service) error {
+		controller := NewWorkItemLinkCategoryController(service, gormapplication.NewGormDB(DB))
+		app.MountWorkItemLinkCategoryController(service, controller)
+		return nil
+	})
 }
