+++ workitem_blackbox_test.go
@@ -4,13 +4,10 @@ import (
 	"bytes"
 	"fmt"
 	"net/http"
-	"net/http/httptest"
 	"strconv"
 	"strings"
 
 	"testing"
-
-	"encoding/json"
 
 	"golang.org/x/net/context"
 
@@ -26,8 +23,6 @@ import (
 	almtoken "github.com/almighty/almighty-core/token"
 	jwt "github.com/dgrijalva/jwt-go"
 	"github.com/goadesign/goa"
-	"github.com/goadesign/goa/middleware"
-	goajwt "github.com/goadesign/goa/middleware/security/jwt"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -169,24 +164,6 @@ func TestListByFields(t *testing.T) {
 	test.DeleteWorkitemOK(t, nil, nil, controller, wi.ID)
 }
 
-// Expected strcture of 401 error response
-type errorResponseStruct struct {
-	Id     string
-	Code   string
-	Status int
-	Detail string
-}
-
-// testSecureAPI defines how a Test object is.
-type testSecureAPI struct {
-	method             string
-	url                string
-	expectedStatusCode int    // this will be tested against responseRecorder.Code
-	expectedErrorCode  string // this will be tested only if response body gets unmarshelled into errorResponseStruct
-	payload            *bytes.Buffer
-	jwtToken           string
-}
-
 func getWorkItemTestData(t *testing.T) []testSecureAPI {
 	privatekey, err := jwt.ParseRSAPrivateKeyFromPEM((configuration.GetTokenPrivateKey()))
 	if err != nil {
@@ -213,28 +190,28 @@ func getWorkItemTestData(t *testing.T) []testSecureAPI {
 		// Create Work Item API with different parameters
 		{
 			method:             http.MethodPost,
-			url:                EndpointWorkItems,
+			url:                endpointWorkItems,
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            createWIPayloadString,
-			jwtToken:           GetExpiredAuthHeader(t, privatekey),
+			jwtToken:           getExpiredAuthHeader(t, privatekey),
 		}, {
 			method:             http.MethodPost,
-			url:                EndpointWorkItems,
+			url:                endpointWorkItems,
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            createWIPayloadString,
-			jwtToken:           GetMalformedAuthHeader(t, privatekey),
+			jwtToken:           getMalformedAuthHeader(t, privatekey),
 		}, {
 			method:             http.MethodPost,
-			url:                EndpointWorkItems,
+			url:                endpointWorkItems,
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            createWIPayloadString,
-			jwtToken:           GetValidAuthHeader(t, differentPrivatekey),
+			jwtToken:           getValidAuthHeader(t, differentPrivatekey),
 		}, {
 			method:             http.MethodPost,
-			url:                EndpointWorkItems,
+			url:                endpointWorkItems,
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            createWIPayloadString,
@@ -243,28 +220,28 @@ func getWorkItemTestData(t *testing.T) []testSecureAPI {
 		// Update Work Item API with different parameters
 		{
 			method:             http.MethodPut,
-			url:                EndpointWorkItems + "/12345",
+			url:                endpointWorkItems + "/12345",
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            createWIPayloadString,
-			jwtToken:           GetExpiredAuthHeader(t, privatekey),
+			jwtToken:           getExpiredAuthHeader(t, privatekey),
 		}, {
 			method:             http.MethodPut,
-			url:                EndpointWorkItems + "/12345",
+			url:                endpointWorkItems + "/12345",
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            createWIPayloadString,
-			jwtToken:           GetMalformedAuthHeader(t, privatekey),
+			jwtToken:           getMalformedAuthHeader(t, privatekey),
 		}, {
 			method:             http.MethodPut,
-			url:                EndpointWorkItems + "/12345",
+			url:                endpointWorkItems + "/12345",
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            createWIPayloadString,
-			jwtToken:           GetValidAuthHeader(t, differentPrivatekey),
+			jwtToken:           getValidAuthHeader(t, differentPrivatekey),
 		}, {
 			method:             http.MethodPut,
-			url:                EndpointWorkItems + "/12345",
+			url:                endpointWorkItems + "/12345",
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            createWIPayloadString,
@@ -273,28 +250,28 @@ func getWorkItemTestData(t *testing.T) []testSecureAPI {
 		// Delete Work Item API with different parameters
 		{
 			method:             http.MethodDelete,
-			url:                EndpointWorkItems + "/12345",
+			url:                endpointWorkItems + "/12345",
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            nil,
-			jwtToken:           GetExpiredAuthHeader(t, privatekey),
+			jwtToken:           getExpiredAuthHeader(t, privatekey),
 		}, {
 			method:             http.MethodDelete,
-			url:                EndpointWorkItems + "/12345",
+			url:                endpointWorkItems + "/12345",
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            nil,
-			jwtToken:           GetMalformedAuthHeader(t, privatekey),
+			jwtToken:           getMalformedAuthHeader(t, privatekey),
 		}, {
 			method:             http.MethodDelete,
-			url:                EndpointWorkItems + "/12345",
+			url:                endpointWorkItems + "/12345",
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            nil,
-			jwtToken:           GetValidAuthHeader(t, differentPrivatekey),
+			jwtToken:           getValidAuthHeader(t, differentPrivatekey),
 		}, {
 			method:             http.MethodDelete,
-			url:                EndpointWorkItems + "/12345",
+			url:                endpointWorkItems + "/12345",
 			expectedStatusCode: http.StatusUnauthorized,
 			expectedErrorCode:  "jwt_security_error",
 			payload:            nil,
@@ -304,7 +281,7 @@ func getWorkItemTestData(t *testing.T) []testSecureAPI {
 		// We do not have security on GET hence this should return 404 not found
 		{
 			method:             http.MethodGet,
-			url:                EndpointWorkItems + "/088481764871",
+			url:                endpointWorkItems + "/088481764871",
 			expectedStatusCode: http.StatusNotFound,
 			expectedErrorCode:  "not_found",
 			payload:            nil,
@@ -315,102 +292,14 @@ func getWorkItemTestData(t *testing.T) []testSecureAPI {
 
 // This test case will check authorized access to Create/Update/Delete APIs
 func TestUnauthorizeWorkItemCUD(t *testing.T) {
-	resource.Require(t, resource.Database)
-
-	// This will be modified after merge PR for "Viper Environment configurations"
-	publickey, err := jwt.ParseRSAPublicKeyFromPEM((configuration.GetTokenPublicKey()))
-	if err != nil {
-		t.Fatal("Could not parse Key ", err)
-	}
-	tokenTests := getWorkItemTestData(t)
-
-	for _, testObject := range tokenTests {
-		// Build a request
-		var req *http.Request
-		var err error
-		if testObject.payload == nil {
-			req, err = http.NewRequest(testObject.method, testObject.url, nil)
-		} else {
-			req, err = http.NewRequest(testObject.method, testObject.url, testObject.payload)
-		}
-		// req, err := http.NewRequest(testObject.method, testObject.url, testObject.payload)
-		if err != nil {
-			t.Fatal("could not create a HTTP request")
-		}
-		// Add Authorization Header
-		req.Header.Add("Authorization", testObject.jwtToken)
-
-		rr := httptest.NewRecorder()
-
-		// temperory service for testing the middleware
-		service := goa.New("TestUnauthorizedCreateWI-Service")
-		assert.NotNil(t, service)
-
-		// if error is thrown during request processing, it will be caught by ErrorHandler middleware
-		// this will put error code, status, details in recorder object.
-		// e.g> {"id":"AL6spYb2","code":"jwt_security_error","status":401,"detail":"JWT validation failed: crypto/rsa: verification error"}
-		service.Use(middleware.ErrorHandler(service, true))
-
-		// append a middleware to service. Use appropriate RSA keys
-		jwtMiddleware := goajwt.New(publickey, nil, app.NewJWTSecurity())
-		// Adding middleware via "app" is important
-		// Because it will check the design and accordingly apply the middleware if mentioned in design
-		// But if I use `service.Use(jwtMiddleware)` then middleware is applied for all the requests (without checking design)
-		app.UseJWTMiddleware(service, jwtMiddleware)
+	UnauthorizeCreateUpdateDeleteTest(t, getWorkItemTestData, func() *goa.Service {
+		return goa.New("TestUnauthorizedCreateWI-Service")
+	}, func(service *goa.Service) error {
 		controller := NewWorkitemController(service, gormapplication.NewGormDB(DB))
 		app.MountWorkitemController(service, controller)
-
-		// Hit the service with own request
-		service.Mux.ServeHTTP(rr, req)
-
-		assert.Equal(t, testObject.expectedStatusCode, rr.Code)
-
-		// Below code tries to open Body response which is expected to be a JSON
-		// If could not parse it correctly into errorResponseStruct
-		// Then it gets logged and continue the test loop
-		content := new(errorResponseStruct)
-		err = json.Unmarshal(rr.Body.Bytes(), content)
-		if err != nil {
-			t.Log("Could not parse JSON response: ", rr.Body)
-			// safe to continue because we alread checked rr.Code=required_value
-			continue
-		}
-		// Additional checks for 'more' confirmation
-		assert.Equal(t, testObject.expectedErrorCode, content.Code)
-		assert.Equal(t, testObject.expectedStatusCode, content.Status)
-	}
+		return nil
+	})
 }
-
-// this key will be used to sign the token but verification should
-// fail as this is not the key used by server security layer
-// ssh-keygen -f test-alm
-var RSADifferentPrivateKeyTest = `-----BEGIN RSA PRIVATE KEY-----
-MIIEpAIBAAKCAQEAsIT76Mr3p8VvtSrzCVcXEcyvalUp50mm4yvfqvZ1fZfbqAzJ
-c4GNJEkpBGoXF+WgjLNkPnwS+k1YuqvPeGG4vFPtErF7nxNCHpzU+cXScOOl3WrM
-S5fj928sBSJiDIDBwc98mQbIKaCrpLSsFe/kapV5mHmmWGAx6dqnObbqtIte4M7w
-arE/c8xW1Fww2YZ4e4Xwknm+Rs2kQmg0SJPpgRih05y3snEQjXz1kR9bGTEBUPmX
-HBTySgA93gmimQUlSAT0+hz9hcYrwCgjbnXUHlcBP2VbB4omJ7L1zJc/XMPwINR/
-PtkGRhL/DXXA4v/0MLkYDXXmZGku/X1+du2ypQIDAQABAoIBAFi0m3si9E2FNFvQ
-l42sDFXPjJ9c6M/n/UvP8niRnf1dYO8Ube/zvJ/tfAVR4wUJSiMqy0dzRn4ufFZi
-nMIcKZ/KdSqdskgAf4uuuIBEXzqHzAR29O9QBymC3pY97xPlaHki8bRc6h2xNlBw
-0sG7agf90btD9soWnT6tuLeSKmRLh5aHUQv3aGwzPyNfKHQ8J/KwKdPudP+tVBsi
-eNd7DZDgSEw6pRaSCKS3ChrsQQ2XPjGo+OI6HjZ/LAFhFXMq2cRGELGF766a6phK
-aCTB619AXiRHdKE98zEY3GMDtXSgeA0yzxcbvr224rEkHGDfkZ0BJwOCqCiaw4tZ
-F/lFDMkCgYEA36Uqyj0cML5rMwC/W4b6ihuK/DujBBFYPQ8eVYt5yUSyLNJn5RLt
-33eBUvgGB/FyAio5aCp49mcPtfFv5GKXpzTSYo/bWV1iy+oVwgPF7UA5gvtRw90w
-NScLNsZ/7fOEpPJvlsKq/PQoMIoAjkegoj95cqM1yzC3aZpaAjx6188CgYEAyg58
-5e5WK3zXICMpE8q+1AB+kJ/3UhQ71kpK4Xml0PtTJ7Bzqn0hiU4ThfpKj1n9PtpU
-9Op3PqcfVjf11SA1tI5LRHQvgUSNppvf2hPgW8QrqRs5YFgNg0DkVXs3OxWIA4QA
-Ko6oZu2ZpvK3TdYXRmcRUXXNyCDoSmJvH339N0sCgYB0g1kCmcm4/0tb+/S1m2Gl
-V+oVtIAeG2csEFdOW+ar27Uzsr5b0nvI4zql3f9OXhR2WkckJJR2UoUV1d3kTxUR
-EGzW2nl9WjChaafCNzMDgmUz/vi/INn/pwKpm8qETkz5njBSi8KHHDBf8VWOynQ+
-cvEzryHUZOH5C2f/KEEbcwKBgQCGzVGgaPjOPJSdWTfPf4T+lXHa9Q4gkWU2WwxI
-D0uD+BiLMxqH1MGqBA/cY5aYutXMuAbT+xUhFIhAkkcNMFcEJaarfcQvvteuHvIi
-YP5e2qqyQHpv/27McV+kc/buEThT+B3QRqqtOLk4+1c1s66Fhr+0FB789I9lCPTQ
-EtL7rwKBgQC5x7lGs+908uqf7yFXHzw7rPGFUe6cuxZ3jVOzovVoXRma+C7nroNx
-/A4rWPqfpmiKcmrd7K4DQFlYhoq+MALEDmQm+/8G6j2inF53fRGgJVzaZhSvnO9X
-CMnDipW5SU9AQE+xC8Zc+02rcyuZ7ha1WXKgIKwAa92jmJSCJjzdxA==
------END RSA PRIVATE KEY-----`
 
 func createPagingTest(t *testing.T, controller *Workitem2Controller, repo *testsupport.WorkItemRepository, totalCount int) func(start int, limit int, first string, last string, prev string, next string) {
 	return func(start int, limit int, first string, last string, prev string, next string) {
